renamed class file to be in line with style guide
